requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
Pillow==8.0.0
pycodestyle==2.6.0
pytz==2020.1
sqlparse==0.4.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
django-environ==0.4.5
flake8==3.8.4
